Select variables

Select a subset of variables from a data frame.

Usage

select2(df, ind)

Arguments

df

A data frame.

ind

Either a vector of variable indices, names, or a boolean vector.

Value

A subsetted data frame.

Examples

select2(iris, 1)
#>     Sepal.Length
#> 1            5.1
#> 2            4.9
#> 3            4.7
#> 4            4.6
#> 5            5.0
#> 6            5.4
#> 7            4.6
#> 8            5.0
#> 9            4.4
#> 10           4.9
#> 11           5.4
#> 12           4.8
#> 13           4.8
#> 14           4.3
#> 15           5.8
#> 16           5.7
#> 17           5.4
#> 18           5.1
#> 19           5.7
#> 20           5.1
#> 21           5.4
#> 22           5.1
#> 23           4.6
#> 24           5.1
#> 25           4.8
#> 26           5.0
#> 27           5.0
#> 28           5.2
#> 29           5.2
#> 30           4.7
#> 31           4.8
#> 32           5.4
#> 33           5.2
#> 34           5.5
#> 35           4.9
#> 36           5.0
#> 37           5.5
#> 38           4.9
#> 39           4.4
#> 40           5.1
#> 41           5.0
#> 42           4.5
#> 43           4.4
#> 44           5.0
#> 45           5.1
#> 46           4.8
#> 47           5.1
#> 48           4.6
#> 49           5.3
#> 50           5.0
#> 51           7.0
#> 52           6.4
#> 53           6.9
#> 54           5.5
#> 55           6.5
#> 56           5.7
#> 57           6.3
#> 58           4.9
#> 59           6.6
#> 60           5.2
#> 61           5.0
#> 62           5.9
#> 63           6.0
#> 64           6.1
#> 65           5.6
#> 66           6.7
#> 67           5.6
#> 68           5.8
#> 69           6.2
#> 70           5.6
#> 71           5.9
#> 72           6.1
#> 73           6.3
#> 74           6.1
#> 75           6.4
#> 76           6.6
#> 77           6.8
#> 78           6.7
#> 79           6.0
#> 80           5.7
#> 81           5.5
#> 82           5.5
#> 83           5.8
#> 84           6.0
#> 85           5.4
#> 86           6.0
#> 87           6.7
#> 88           6.3
#> 89           5.6
#> 90           5.5
#> 91           5.5
#> 92           6.1
#> 93           5.8
#> 94           5.0
#> 95           5.6
#> 96           5.7
#> 97           5.7
#> 98           6.2
#> 99           5.1
#> 100          5.7
#> 101          6.3
#> 102          5.8
#> 103          7.1
#> 104          6.3
#> 105          6.5
#> 106          7.6
#> 107          4.9
#> 108          7.3
#> 109          6.7
#> 110          7.2
#> 111          6.5
#> 112          6.4
#> 113          6.8
#> 114          5.7
#> 115          5.8
#> 116          6.4
#> 117          6.5
#> 118          7.7
#> 119          7.7
#> 120          6.0
#> 121          6.9
#> 122          5.6
#> 123          7.7
#> 124          6.3
#> 125          6.7
#> 126          7.2
#> 127          6.2
#> 128          6.1
#> 129          6.4
#> 130          7.2
#> 131          7.4
#> 132          7.9
#> 133          6.4
#> 134          6.3
#> 135          6.1
#> 136          7.7
#> 137          6.3
#> 138          6.4
#> 139          6.0
#> 140          6.9
#> 141          6.7
#> 142          6.9
#> 143          5.8
#> 144          6.8
#> 145          6.7
#> 146          6.7
#> 147          6.3
#> 148          6.5
#> 149          6.2
#> 150          5.9
select2(iris, 1:3)
#>     Sepal.Length Sepal.Width Petal.Length
#> 1            5.1         3.5          1.4
#> 2            4.9         3.0          1.4
#> 3            4.7         3.2          1.3
#> 4            4.6         3.1          1.5
#> 5            5.0         3.6          1.4
#> 6            5.4         3.9          1.7
#> 7            4.6         3.4          1.4
#> 8            5.0         3.4          1.5
#> 9            4.4         2.9          1.4
#> 10           4.9         3.1          1.5
#> 11           5.4         3.7          1.5
#> 12           4.8         3.4          1.6
#> 13           4.8         3.0          1.4
#> 14           4.3         3.0          1.1
#> 15           5.8         4.0          1.2
#> 16           5.7         4.4          1.5
#> 17           5.4         3.9          1.3
#> 18           5.1         3.5          1.4
#> 19           5.7         3.8          1.7
#> 20           5.1         3.8          1.5
#> 21           5.4         3.4          1.7
#> 22           5.1         3.7          1.5
#> 23           4.6         3.6          1.0
#> 24           5.1         3.3          1.7
#> 25           4.8         3.4          1.9
#> 26           5.0         3.0          1.6
#> 27           5.0         3.4          1.6
#> 28           5.2         3.5          1.5
#> 29           5.2         3.4          1.4
#> 30           4.7         3.2          1.6
#> 31           4.8         3.1          1.6
#> 32           5.4         3.4          1.5
#> 33           5.2         4.1          1.5
#> 34           5.5         4.2          1.4
#> 35           4.9         3.1          1.5
#> 36           5.0         3.2          1.2
#> 37           5.5         3.5          1.3
#> 38           4.9         3.6          1.4
#> 39           4.4         3.0          1.3
#> 40           5.1         3.4          1.5
#> 41           5.0         3.5          1.3
#> 42           4.5         2.3          1.3
#> 43           4.4         3.2          1.3
#> 44           5.0         3.5          1.6
#> 45           5.1         3.8          1.9
#> 46           4.8         3.0          1.4
#> 47           5.1         3.8          1.6
#> 48           4.6         3.2          1.4
#> 49           5.3         3.7          1.5
#> 50           5.0         3.3          1.4
#> 51           7.0         3.2          4.7
#> 52           6.4         3.2          4.5
#> 53           6.9         3.1          4.9
#> 54           5.5         2.3          4.0
#> 55           6.5         2.8          4.6
#> 56           5.7         2.8          4.5
#> 57           6.3         3.3          4.7
#> 58           4.9         2.4          3.3
#> 59           6.6         2.9          4.6
#> 60           5.2         2.7          3.9
#> 61           5.0         2.0          3.5
#> 62           5.9         3.0          4.2
#> 63           6.0         2.2          4.0
#> 64           6.1         2.9          4.7
#> 65           5.6         2.9          3.6
#> 66           6.7         3.1          4.4
#> 67           5.6         3.0          4.5
#> 68           5.8         2.7          4.1
#> 69           6.2         2.2          4.5
#> 70           5.6         2.5          3.9
#> 71           5.9         3.2          4.8
#> 72           6.1         2.8          4.0
#> 73           6.3         2.5          4.9
#> 74           6.1         2.8          4.7
#> 75           6.4         2.9          4.3
#> 76           6.6         3.0          4.4
#> 77           6.8         2.8          4.8
#> 78           6.7         3.0          5.0
#> 79           6.0         2.9          4.5
#> 80           5.7         2.6          3.5
#> 81           5.5         2.4          3.8
#> 82           5.5         2.4          3.7
#> 83           5.8         2.7          3.9
#> 84           6.0         2.7          5.1
#> 85           5.4         3.0          4.5
#> 86           6.0         3.4          4.5
#> 87           6.7         3.1          4.7
#> 88           6.3         2.3          4.4
#> 89           5.6         3.0          4.1
#> 90           5.5         2.5          4.0
#> 91           5.5         2.6          4.4
#> 92           6.1         3.0          4.6
#> 93           5.8         2.6          4.0
#> 94           5.0         2.3          3.3
#> 95           5.6         2.7          4.2
#> 96           5.7         3.0          4.2
#> 97           5.7         2.9          4.2
#> 98           6.2         2.9          4.3
#> 99           5.1         2.5          3.0
#> 100          5.7         2.8          4.1
#> 101          6.3         3.3          6.0
#> 102          5.8         2.7          5.1
#> 103          7.1         3.0          5.9
#> 104          6.3         2.9          5.6
#> 105          6.5         3.0          5.8
#> 106          7.6         3.0          6.6
#> 107          4.9         2.5          4.5
#> 108          7.3         2.9          6.3
#> 109          6.7         2.5          5.8
#> 110          7.2         3.6          6.1
#> 111          6.5         3.2          5.1
#> 112          6.4         2.7          5.3
#> 113          6.8         3.0          5.5
#> 114          5.7         2.5          5.0
#> 115          5.8         2.8          5.1
#> 116          6.4         3.2          5.3
#> 117          6.5         3.0          5.5
#> 118          7.7         3.8          6.7
#> 119          7.7         2.6          6.9
#> 120          6.0         2.2          5.0
#> 121          6.9         3.2          5.7
#> 122          5.6         2.8          4.9
#> 123          7.7         2.8          6.7
#> 124          6.3         2.7          4.9
#> 125          6.7         3.3          5.7
#> 126          7.2         3.2          6.0
#> 127          6.2         2.8          4.8
#> 128          6.1         3.0          4.9
#> 129          6.4         2.8          5.6
#> 130          7.2         3.0          5.8
#> 131          7.4         2.8          6.1
#> 132          7.9         3.8          6.4
#> 133          6.4         2.8          5.6
#> 134          6.3         2.8          5.1
#> 135          6.1         2.6          5.6
#> 136          7.7         3.0          6.1
#> 137          6.3         3.4          5.6
#> 138          6.4         3.1          5.5
#> 139          6.0         3.0          4.8
#> 140          6.9         3.1          5.4
#> 141          6.7         3.1          5.6
#> 142          6.9         3.1          5.1
#> 143          5.8         2.7          5.1
#> 144          6.8         3.2          5.9
#> 145          6.7         3.3          5.7
#> 146          6.7         3.0          5.2
#> 147          6.3         2.5          5.0
#> 148          6.5         3.0          5.2
#> 149          6.2         3.4          5.4
#> 150          5.9         3.0          5.1
select2(iris, -1)
#>     Sepal.Width Petal.Length Petal.Width    Species
#> 1           3.5          1.4         0.2     setosa
#> 2           3.0          1.4         0.2     setosa
#> 3           3.2          1.3         0.2     setosa
#> 4           3.1          1.5         0.2     setosa
#> 5           3.6          1.4         0.2     setosa
#> 6           3.9          1.7         0.4     setosa
#> 7           3.4          1.4         0.3     setosa
#> 8           3.4          1.5         0.2     setosa
#> 9           2.9          1.4         0.2     setosa
#> 10          3.1          1.5         0.1     setosa
#> 11          3.7          1.5         0.2     setosa
#> 12          3.4          1.6         0.2     setosa
#> 13          3.0          1.4         0.1     setosa
#> 14          3.0          1.1         0.1     setosa
#> 15          4.0          1.2         0.2     setosa
#> 16          4.4          1.5         0.4     setosa
#> 17          3.9          1.3         0.4     setosa
#> 18          3.5          1.4         0.3     setosa
#> 19          3.8          1.7         0.3     setosa
#> 20          3.8          1.5         0.3     setosa
#> 21          3.4          1.7         0.2     setosa
#> 22          3.7          1.5         0.4     setosa
#> 23          3.6          1.0         0.2     setosa
#> 24          3.3          1.7         0.5     setosa
#> 25          3.4          1.9         0.2     setosa
#> 26          3.0          1.6         0.2     setosa
#> 27          3.4          1.6         0.4     setosa
#> 28          3.5          1.5         0.2     setosa
#> 29          3.4          1.4         0.2     setosa
#> 30          3.2          1.6         0.2     setosa
#> 31          3.1          1.6         0.2     setosa
#> 32          3.4          1.5         0.4     setosa
#> 33          4.1          1.5         0.1     setosa
#> 34          4.2          1.4         0.2     setosa
#> 35          3.1          1.5         0.2     setosa
#> 36          3.2          1.2         0.2     setosa
#> 37          3.5          1.3         0.2     setosa
#> 38          3.6          1.4         0.1     setosa
#> 39          3.0          1.3         0.2     setosa
#> 40          3.4          1.5         0.2     setosa
#> 41          3.5          1.3         0.3     setosa
#> 42          2.3          1.3         0.3     setosa
#> 43          3.2          1.3         0.2     setosa
#> 44          3.5          1.6         0.6     setosa
#> 45          3.8          1.9         0.4     setosa
#> 46          3.0          1.4         0.3     setosa
#> 47          3.8          1.6         0.2     setosa
#> 48          3.2          1.4         0.2     setosa
#> 49          3.7          1.5         0.2     setosa
#> 50          3.3          1.4         0.2     setosa
#> 51          3.2          4.7         1.4 versicolor
#> 52          3.2          4.5         1.5 versicolor
#> 53          3.1          4.9         1.5 versicolor
#> 54          2.3          4.0         1.3 versicolor
#> 55          2.8          4.6         1.5 versicolor
#> 56          2.8          4.5         1.3 versicolor
#> 57          3.3          4.7         1.6 versicolor
#> 58          2.4          3.3         1.0 versicolor
#> 59          2.9          4.6         1.3 versicolor
#> 60          2.7          3.9         1.4 versicolor
#> 61          2.0          3.5         1.0 versicolor
#> 62          3.0          4.2         1.5 versicolor
#> 63          2.2          4.0         1.0 versicolor
#> 64          2.9          4.7         1.4 versicolor
#> 65          2.9          3.6         1.3 versicolor
#> 66          3.1          4.4         1.4 versicolor
#> 67          3.0          4.5         1.5 versicolor
#> 68          2.7          4.1         1.0 versicolor
#> 69          2.2          4.5         1.5 versicolor
#> 70          2.5          3.9         1.1 versicolor
#> 71          3.2          4.8         1.8 versicolor
#> 72          2.8          4.0         1.3 versicolor
#> 73          2.5          4.9         1.5 versicolor
#> 74          2.8          4.7         1.2 versicolor
#> 75          2.9          4.3         1.3 versicolor
#> 76          3.0          4.4         1.4 versicolor
#> 77          2.8          4.8         1.4 versicolor
#> 78          3.0          5.0         1.7 versicolor
#> 79          2.9          4.5         1.5 versicolor
#> 80          2.6          3.5         1.0 versicolor
#> 81          2.4          3.8         1.1 versicolor
#> 82          2.4          3.7         1.0 versicolor
#> 83          2.7          3.9         1.2 versicolor
#> 84          2.7          5.1         1.6 versicolor
#> 85          3.0          4.5         1.5 versicolor
#> 86          3.4          4.5         1.6 versicolor
#> 87          3.1          4.7         1.5 versicolor
#> 88          2.3          4.4         1.3 versicolor
#> 89          3.0          4.1         1.3 versicolor
#> 90          2.5          4.0         1.3 versicolor
#> 91          2.6          4.4         1.2 versicolor
#> 92          3.0          4.6         1.4 versicolor
#> 93          2.6          4.0         1.2 versicolor
#> 94          2.3          3.3         1.0 versicolor
#> 95          2.7          4.2         1.3 versicolor
#> 96          3.0          4.2         1.2 versicolor
#> 97          2.9          4.2         1.3 versicolor
#> 98          2.9          4.3         1.3 versicolor
#> 99          2.5          3.0         1.1 versicolor
#> 100         2.8          4.1         1.3 versicolor
#> 101         3.3          6.0         2.5  virginica
#> 102         2.7          5.1         1.9  virginica
#> 103         3.0          5.9         2.1  virginica
#> 104         2.9          5.6         1.8  virginica
#> 105         3.0          5.8         2.2  virginica
#> 106         3.0          6.6         2.1  virginica
#> 107         2.5          4.5         1.7  virginica
#> 108         2.9          6.3         1.8  virginica
#> 109         2.5          5.8         1.8  virginica
#> 110         3.6          6.1         2.5  virginica
#> 111         3.2          5.1         2.0  virginica
#> 112         2.7          5.3         1.9  virginica
#> 113         3.0          5.5         2.1  virginica
#> 114         2.5          5.0         2.0  virginica
#> 115         2.8          5.1         2.4  virginica
#> 116         3.2          5.3         2.3  virginica
#> 117         3.0          5.5         1.8  virginica
#> 118         3.8          6.7         2.2  virginica
#> 119         2.6          6.9         2.3  virginica
#> 120         2.2          5.0         1.5  virginica
#> 121         3.2          5.7         2.3  virginica
#> 122         2.8          4.9         2.0  virginica
#> 123         2.8          6.7         2.0  virginica
#> 124         2.7          4.9         1.8  virginica
#> 125         3.3          5.7         2.1  virginica
#> 126         3.2          6.0         1.8  virginica
#> 127         2.8          4.8         1.8  virginica
#> 128         3.0          4.9         1.8  virginica
#> 129         2.8          5.6         2.1  virginica
#> 130         3.0          5.8         1.6  virginica
#> 131         2.8          6.1         1.9  virginica
#> 132         3.8          6.4         2.0  virginica
#> 133         2.8          5.6         2.2  virginica
#> 134         2.8          5.1         1.5  virginica
#> 135         2.6          5.6         1.4  virginica
#> 136         3.0          6.1         2.3  virginica
#> 137         3.4          5.6         2.4  virginica
#> 138         3.1          5.5         1.8  virginica
#> 139         3.0          4.8         1.8  virginica
#> 140         3.1          5.4         2.1  virginica
#> 141         3.1          5.6         2.4  virginica
#> 142         3.1          5.1         2.3  virginica
#> 143         2.7          5.1         1.9  virginica
#> 144         3.2          5.9         2.3  virginica
#> 145         3.3          5.7         2.5  virginica
#> 146         3.0          5.2         2.3  virginica
#> 147         2.5          5.0         1.9  virginica
#> 148         3.0          5.2         2.0  virginica
#> 149         3.4          5.4         2.3  virginica
#> 150         3.0          5.1         1.8  virginica
select2(iris, c("Sepal.Width", "Species"))
#>     Sepal.Width    Species
#> 1           3.5     setosa
#> 2           3.0     setosa
#> 3           3.2     setosa
#> 4           3.1     setosa
#> 5           3.6     setosa
#> 6           3.9     setosa
#> 7           3.4     setosa
#> 8           3.4     setosa
#> 9           2.9     setosa
#> 10          3.1     setosa
#> 11          3.7     setosa
#> 12          3.4     setosa
#> 13          3.0     setosa
#> 14          3.0     setosa
#> 15          4.0     setosa
#> 16          4.4     setosa
#> 17          3.9     setosa
#> 18          3.5     setosa
#> 19          3.8     setosa
#> 20          3.8     setosa
#> 21          3.4     setosa
#> 22          3.7     setosa
#> 23          3.6     setosa
#> 24          3.3     setosa
#> 25          3.4     setosa
#> 26          3.0     setosa
#> 27          3.4     setosa
#> 28          3.5     setosa
#> 29          3.4     setosa
#> 30          3.2     setosa
#> 31          3.1     setosa
#> 32          3.4     setosa
#> 33          4.1     setosa
#> 34          4.2     setosa
#> 35          3.1     setosa
#> 36          3.2     setosa
#> 37          3.5     setosa
#> 38          3.6     setosa
#> 39          3.0     setosa
#> 40          3.4     setosa
#> 41          3.5     setosa
#> 42          2.3     setosa
#> 43          3.2     setosa
#> 44          3.5     setosa
#> 45          3.8     setosa
#> 46          3.0     setosa
#> 47          3.8     setosa
#> 48          3.2     setosa
#> 49          3.7     setosa
#> 50          3.3     setosa
#> 51          3.2 versicolor
#> 52          3.2 versicolor
#> 53          3.1 versicolor
#> 54          2.3 versicolor
#> 55          2.8 versicolor
#> 56          2.8 versicolor
#> 57          3.3 versicolor
#> 58          2.4 versicolor
#> 59          2.9 versicolor
#> 60          2.7 versicolor
#> 61          2.0 versicolor
#> 62          3.0 versicolor
#> 63          2.2 versicolor
#> 64          2.9 versicolor
#> 65          2.9 versicolor
#> 66          3.1 versicolor
#> 67          3.0 versicolor
#> 68          2.7 versicolor
#> 69          2.2 versicolor
#> 70          2.5 versicolor
#> 71          3.2 versicolor
#> 72          2.8 versicolor
#> 73          2.5 versicolor
#> 74          2.8 versicolor
#> 75          2.9 versicolor
#> 76          3.0 versicolor
#> 77          2.8 versicolor
#> 78          3.0 versicolor
#> 79          2.9 versicolor
#> 80          2.6 versicolor
#> 81          2.4 versicolor
#> 82          2.4 versicolor
#> 83          2.7 versicolor
#> 84          2.7 versicolor
#> 85          3.0 versicolor
#> 86          3.4 versicolor
#> 87          3.1 versicolor
#> 88          2.3 versicolor
#> 89          3.0 versicolor
#> 90          2.5 versicolor
#> 91          2.6 versicolor
#> 92          3.0 versicolor
#> 93          2.6 versicolor
#> 94          2.3 versicolor
#> 95          2.7 versicolor
#> 96          3.0 versicolor
#> 97          2.9 versicolor
#> 98          2.9 versicolor
#> 99          2.5 versicolor
#> 100         2.8 versicolor
#> 101         3.3  virginica
#> 102         2.7  virginica
#> 103         3.0  virginica
#> 104         2.9  virginica
#> 105         3.0  virginica
#> 106         3.0  virginica
#> 107         2.5  virginica
#> 108         2.9  virginica
#> 109         2.5  virginica
#> 110         3.6  virginica
#> 111         3.2  virginica
#> 112         2.7  virginica
#> 113         3.0  virginica
#> 114         2.5  virginica
#> 115         2.8  virginica
#> 116         3.2  virginica
#> 117         3.0  virginica
#> 118         3.8  virginica
#> 119         2.6  virginica
#> 120         2.2  virginica
#> 121         3.2  virginica
#> 122         2.8  virginica
#> 123         2.8  virginica
#> 124         2.7  virginica
#> 125         3.3  virginica
#> 126         3.2  virginica
#> 127         2.8  virginica
#> 128         3.0  virginica
#> 129         2.8  virginica
#> 130         3.0  virginica
#> 131         2.8  virginica
#> 132         3.8  virginica
#> 133         2.8  virginica
#> 134         2.8  virginica
#> 135         2.6  virginica
#> 136         3.0  virginica
#> 137         3.4  virginica
#> 138         3.1  virginica
#> 139         3.0  virginica
#> 140         3.1  virginica
#> 141         3.1  virginica
#> 142         3.1  virginica
#> 143         2.7  virginica
#> 144         3.2  virginica
#> 145         3.3  virginica
#> 146         3.0  virginica
#> 147         2.5  virginica
#> 148         3.0  virginica
#> 149         3.4  virginica
#> 150         3.0  virginica
select2(iris, c(TRUE, FALSE, FALSE, TRUE, FALSE))
#>     Sepal.Length Petal.Width
#> 1            5.1         0.2
#> 2            4.9         0.2
#> 3            4.7         0.2
#> 4            4.6         0.2
#> 5            5.0         0.2
#> 6            5.4         0.4
#> 7            4.6         0.3
#> 8            5.0         0.2
#> 9            4.4         0.2
#> 10           4.9         0.1
#> 11           5.4         0.2
#> 12           4.8         0.2
#> 13           4.8         0.1
#> 14           4.3         0.1
#> 15           5.8         0.2
#> 16           5.7         0.4
#> 17           5.4         0.4
#> 18           5.1         0.3
#> 19           5.7         0.3
#> 20           5.1         0.3
#> 21           5.4         0.2
#> 22           5.1         0.4
#> 23           4.6         0.2
#> 24           5.1         0.5
#> 25           4.8         0.2
#> 26           5.0         0.2
#> 27           5.0         0.4
#> 28           5.2         0.2
#> 29           5.2         0.2
#> 30           4.7         0.2
#> 31           4.8         0.2
#> 32           5.4         0.4
#> 33           5.2         0.1
#> 34           5.5         0.2
#> 35           4.9         0.2
#> 36           5.0         0.2
#> 37           5.5         0.2
#> 38           4.9         0.1
#> 39           4.4         0.2
#> 40           5.1         0.2
#> 41           5.0         0.3
#> 42           4.5         0.3
#> 43           4.4         0.2
#> 44           5.0         0.6
#> 45           5.1         0.4
#> 46           4.8         0.3
#> 47           5.1         0.2
#> 48           4.6         0.2
#> 49           5.3         0.2
#> 50           5.0         0.2
#> 51           7.0         1.4
#> 52           6.4         1.5
#> 53           6.9         1.5
#> 54           5.5         1.3
#> 55           6.5         1.5
#> 56           5.7         1.3
#> 57           6.3         1.6
#> 58           4.9         1.0
#> 59           6.6         1.3
#> 60           5.2         1.4
#> 61           5.0         1.0
#> 62           5.9         1.5
#> 63           6.0         1.0
#> 64           6.1         1.4
#> 65           5.6         1.3
#> 66           6.7         1.4
#> 67           5.6         1.5
#> 68           5.8         1.0
#> 69           6.2         1.5
#> 70           5.6         1.1
#> 71           5.9         1.8
#> 72           6.1         1.3
#> 73           6.3         1.5
#> 74           6.1         1.2
#> 75           6.4         1.3
#> 76           6.6         1.4
#> 77           6.8         1.4
#> 78           6.7         1.7
#> 79           6.0         1.5
#> 80           5.7         1.0
#> 81           5.5         1.1
#> 82           5.5         1.0
#> 83           5.8         1.2
#> 84           6.0         1.6
#> 85           5.4         1.5
#> 86           6.0         1.6
#> 87           6.7         1.5
#> 88           6.3         1.3
#> 89           5.6         1.3
#> 90           5.5         1.3
#> 91           5.5         1.2
#> 92           6.1         1.4
#> 93           5.8         1.2
#> 94           5.0         1.0
#> 95           5.6         1.3
#> 96           5.7         1.2
#> 97           5.7         1.3
#> 98           6.2         1.3
#> 99           5.1         1.1
#> 100          5.7         1.3
#> 101          6.3         2.5
#> 102          5.8         1.9
#> 103          7.1         2.1
#> 104          6.3         1.8
#> 105          6.5         2.2
#> 106          7.6         2.1
#> 107          4.9         1.7
#> 108          7.3         1.8
#> 109          6.7         1.8
#> 110          7.2         2.5
#> 111          6.5         2.0
#> 112          6.4         1.9
#> 113          6.8         2.1
#> 114          5.7         2.0
#> 115          5.8         2.4
#> 116          6.4         2.3
#> 117          6.5         1.8
#> 118          7.7         2.2
#> 119          7.7         2.3
#> 120          6.0         1.5
#> 121          6.9         2.3
#> 122          5.6         2.0
#> 123          7.7         2.0
#> 124          6.3         1.8
#> 125          6.7         2.1
#> 126          7.2         1.8
#> 127          6.2         1.8
#> 128          6.1         1.8
#> 129          6.4         2.1
#> 130          7.2         1.6
#> 131          7.4         1.9
#> 132          7.9         2.0
#> 133          6.4         2.2
#> 134          6.3         1.5
#> 135          6.1         1.4
#> 136          7.7         2.3
#> 137          6.3         2.4
#> 138          6.4         1.8
#> 139          6.0         1.8
#> 140          6.9         2.1
#> 141          6.7         2.4
#> 142          6.9         2.3
#> 143          5.8         1.9
#> 144          6.8         2.3
#> 145          6.7         2.5
#> 146          6.7         2.3
#> 147          6.3         1.9
#> 148          6.5         2.0
#> 149          6.2         2.3
#> 150          5.9         1.8